fix migrate error type

+++ database/migrations/2019_02_06_204006_create_table_results.php
@@ -17,7 +17,7 @@ class CreateTableResults extends Migration
             $table->increments('id');
             $table->string('email',255);
             $table->integer('grades');
-            $table->float('total_time');
+            $table->integer('total_time');
             $table->timestamps();
         });
     }
